fix(db): fresh-replay guards for three already-applied migrations (Supabase Preview)
A from-scratch replay of the chain (what Supabase Preview does) still failed
after the strategy_filter_analysis fix; each of these versions is recorded on
prod, so the edits are Preview-only:

- 20260108193936: ALTER TABLE exchange_credentials ADD COLUMN IF NOT EXISTS
  user_id UUID NOT NULL before its policies — 20251013180000 created the table
  without user_id; prod has user_id uuid NOT NULL (verified read-only).
- 20260202190000: ::NUMERIC cast inside ROUND() for profit_factor —
  round(double precision, int) does not exist; same cast the follow-up
  reconciliation 20260202192540 uses.
- 20260202192540: DROP VIEW IF EXISTS daily_trade_summary before CREATE —
  CREATE OR REPLACE VIEW cannot retype columns left by 20260202190000; prod
  was a fresh create here.

Full-chain replay on PostgreSQL 16: all 57 files apply cleanly.

+++ supabase/migrations/20260202190000_fix_trade_analytics_schema.sql
@@ -122,8 +122,11 @@ SELECT
   END as win_rate,
   CASE
     WHEN COALESCE(SUM(ABS(realized_pnl)) FILTER (WHERE outcome = 'loss'), 0) > 0
-    THEN ROUND(COALESCE(SUM(realized_pnl) FILTER (WHERE outcome = 'win'), 0) /
-         COALESCE(SUM(ABS(realized_pnl)) FILTER (WHERE outcome = 'loss'), 1), 2)
+    -- Fresh-replay guard (2026-09-10, Supabase Preview): realized_pnl is DOUBLE
+    -- PRECISION here and round(double precision, int) does not exist; cast to
+    -- NUMERIC exactly as the follow-up reconciliation 20260202192540 does.
+    THEN ROUND((COALESCE(SUM(realized_pnl) FILTER (WHERE outcome = 'win'), 0) /
+         COALESCE(SUM(ABS(realized_pnl)) FILTER (WHERE outcome = 'loss'), 1))::NUMERIC, 2)
     ELSE NULL
   END as profit_factor,
   ROUND(AVG(duration_seconds)::NUMERIC, 0) as avg_duration_seconds,
